Factor out common CreateException and GetException logic
