+++ src/Exception/MachineProvider/GetException.php
@@ -4,23 +4,17 @@ namespace App\Exception\MachineProvider;
 
 use App\Entity\Worker;
 
-class GetException extends AbstractRemoteApiWrappingException implements RemoteApiExceptionWrapperInterface
+class GetException extends AbstractWorkerApiActionException
 {
     private const MESSAGE = 'Unable to get remote machine for worker %s %s';
 
-    public function __construct(
-        private Worker $worker,
-        private \Throwable $remoteApiException
-    ) {
+    public function __construct(Worker $worker, \Throwable $remoteApiException)
+    {
         parent::__construct(
             sprintf(self::MESSAGE, (string) $worker, $worker->getLabel()),
             0,
+            $worker,
             $remoteApiException
         );
-    }
-
-    public function getWorker(): Worker
-    {
-        return $this->worker;
     }
 }
